src/requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==3.0.2
Flask[async]
gunicorn==20.1.0
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
shellingham==1.4.0
typer==0.7.0
typer-cli==0.0.13
typing_extensions==4.5.0
Werkzeug==3.0.1
zipp==3.15.0
requests==2.31.0
python-dotenv==1.0.0
